refactor: decompose local video generation service (#4892)
Extract render argument assembly, process supervision, frame handling, stitching, upscaling, history mutations, and chained generation into focused modules while preserving the local.js public facade.
